Abstracted Server Initialization and Termination
i moved the function woohoo winter break
