+++ test/Event.js
@@ -6,9 +6,8 @@ const User = require('../api/models/User')
 // Require the dev-dependencies
 const chai = require('chai')
 const chaiHttp = require('chai-http')
-const server = require('../server')
-let serverInstance = null
 let app = null
+
 const expect = chai.expect
 // tools for testing
 const tools = require('../util/testing-utils/tools.js')
@@ -16,26 +15,16 @@ const tools = require('../util/testing-utils/tools.js')
 chai.should()
 chai.use(chaiHttp)
 
-function initializeServer () {
-  serverInstance = new server.Server()
-  serverInstance.openConnection()
-  app = serverInstance.getServerInstance()
-}
-
-function terminateServer (done) {
-  serverInstance.closeConnection(done)
-}
-
 describe('Event', () => {
   before(done => {
-    initializeServer()
+    app = tools.initializeServer()
     // Before each test we empty the database
     tools.emptySchema(Event)
     tools.emptySchema(User)
     done()
   })
   after(done => {
-    terminateServer(done)
+    tools.terminateServer(done)
   })
 
   let token = ''
