+++ test/user.js
@@ -7,8 +7,7 @@ const User = require('../api/models/User')
 // Require the dev-dependencies
 const chai = require('chai')
 const chaiHttp = require('chai-http')
-const server = require('../server')
-let serverInstance = null
+
 let app = null
 const expect = chai.expect
 // tools for testing
@@ -17,26 +16,16 @@ const tools = require('../util/testing-utils/tools.js')
 chai.should()
 chai.use(chaiHttp)
 
-function initializeServer () {
-  serverInstance = new server.Server()
-  serverInstance.openConnection()
-  app = serverInstance.getServerInstance()
-}
-
-function terminateServer (done) {
-  serverInstance.closeConnection(done)
-}
-
 // Our parent block
 describe('Users', () => {
   before(done => {
-    initializeServer()
+    app = tools.initializeServer()
     // Before each test we empty the database
     tools.emptySchema(User)
     done()
   })
   after(done => {
-    terminateServer(done)
+    tools.terminateServer(done)
   })
 
   let token = ''
