+++ util/testing-utils/tools.js
@@ -6,7 +6,22 @@ function emptySchema (schema) {
   })
 }
 
+const server = require('../../server')
+let serverInstance = null
+
+function initializeServer () {
+  serverInstance = new server.Server()
+  serverInstance.openConnection()
+  return serverInstance.getServerInstance()
+}
+
+function terminateServer (done) {
+  serverInstance.closeConnection(done)
+}
+
 // Exporting functions
 module.exports = {
-  emptySchema: emptySchema
+  emptySchema: emptySchema,
+  initializeServer: initializeServer,
+  terminateServer: terminateServer
 }
